feat: move group by month and year to extensions file
